+++ v3/pipeline.go
@@ -81,7 +81,7 @@ func PipelineController(sq *Sequence) Controller {
 
 		// special case - do we have a non-zero status code, but no error?
 		if sq.StatusCode != StatusOkay && sq.Err == nil {
-			sq.Err = ErrPipelineNonZeroStatusCode{sq.StatusCode}
+			sq.Err = ErrNonZeroStatusCode{"pipeline", sq.StatusCode}
 		}
 
 		// all done